Add basic date display on right of package list
Not sure this is the best location but it saves on the hassle of trying
to get them to neatly line up with the headers which is tricky due to
the different font sizes and maybe the underline.

Currently displays an ugly iso8601 format for the date. Would be better
as '3 days ago' style display but I'm not sure of the best package/lib
to help with that.